feature: 提交

+++ giphy.py
@@ -4,6 +4,7 @@ import json
 import os
 import requests
 
+
 class ImageOriginal:
     def __init__(self, url, webp):
         self.url = url
@@ -16,13 +17,14 @@ class Images:
 
 
 class Result:
-    def __init__(self, title, images: Images):
+    def __init__(self, id_str, title, images: Images):
+        self.id_str = id_str
         self.title = title
         self.images = images
 
 
 class TagResponse:
-    def __init__(self, next_url, results):
+    def __init__(self, next_url, results: list):
         self.next_url = next_url
         self.results = results
 
@@ -45,12 +47,13 @@ def parse_image_tag(tag_str):
     results = []
     for result_dict in results_dict:
         # create a list of Image objects from the 'images' di21ctionary
+        id_str = result_dict.get('id')
         title = result_dict.get('title')
         images_dict = result_dict.get('images')
         original_dict = images_dict.get('original')
         original = ImageOriginal(original_dict['url'], original_dict['webp'])
         images = Images(original)
-        results.append(Result(title, images))
+        results.append(Result(id_str, title, images))
     # create a new TagResponse object with the parsed data
     tag_response = TagResponse(next_url, results)
     return tag_response
@@ -58,8 +61,8 @@ def parse_image_tag(tag_str):
 
 def download_result(result: Result, parent_fold):
     if result.images.original:
-        download_image(result.images.original.url, os.path.join(parent_fold, 'original'))
-        download_image(result.images.original.webp, os.path.join(parent_fold, 'original'))
+        download_image(result.images.original.url, os.path.join(parent_fold, result.id_str, 'original'))
+        download_image(result.images.original.webp, os.path.join(parent_fold, result.id_str, 'original'))
 
 
 def download_tag_image(response: TagResponse, parent_fold):
@@ -68,9 +71,8 @@ def download_tag_image(response: TagResponse, parent_fold):
 
 
 def download_image(url, parent_fold):
+    print('download_image ------- start ', url)
     os.makedirs(parent_fold, exist_ok=True)
-    # Make a request to the URL to get the image data.
-    response = requests.get(url)
 
     # Get the file extension from the URL.
     file_extension = url.split('.')[-1]
@@ -81,17 +83,45 @@ def download_image(url, parent_fold):
     # Combine the parent folder path and filename to create the full file path.
     file_path = os.path.join(parent_fold, filename)
 
-    # Write the image data to the file.
-    with open(file_path, 'wb') as file:
-        file.write(response.content)
+    if os.path.exists(file_path):
+        print(f"Image already exists locally: {file_path}")
+    else:
+        # Make a request to the URL to get the image data.
+        response = requests.get(url)
+        # Write the image data to the file.
+        with open(file_path, 'wb') as file:
+            file.write(response.content)
+
+    print('download_image ------- end ', url)
+
+
+def download_tag_2(url, parent_fold):
+    print('download_tag_2 start ', url)
+    response_str = request_url(url)
+    response = parse_image_tag(response_str)
+    download_tag_image(response, parent_fold)
+    print('download_tag_2 end ', url)
+    return response
+
+
+max_count = 100
+
+
+def download_tag(url, parent_fold):
+    print('download_tag start ', url)
+    count = 0
+    i = 0
+    while i < 10:
+        response = download_tag_2(url, parent_fold)
+        count = count + len(response.results)
+        if response is None or response.next_url is None or count > max_count:
+            break
+        i += 1
+    print('download_tag end ', url)
 
 
 def main():
-    response_str = request_url('https://giphy.com/api/v4/channels/11313767/feed/?offset=0')
-    result = parse_image_tag(response_str)
-    download_tag_image(result, './gif')
-    print(result)
+    download_tag('https://giphy.com/api/v4/channels/11313767/feed/?offset=0', "./gif")
 
 
 main()
-
